Fix Win64 compile of videoadapter_unittest.cc.
Missed an typecast in videoadapter_unittest.cc in r6979 due to
tryservers being clogged and me waiting for a windows, linux, mac and
tsanv2 bot to finish was not enough. Committing fix straight away to
un-break tree.

TBR=tommi@webrtc.org
BUG=3671

Review URL: https://webrtc-codereview.appspot.com/18279004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@6980 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ talk/media/base/videoadapter_unittest.cc
@@ -121,8 +121,8 @@ class VideoAdapterTest : public testing::Test {
       stats.dropped_frames = dropped_frames_;
       stats.last_adapt_was_no_op = last_adapt_was_no_op_;
       if (adapted_frame_ != NULL) {
-        stats.adapted_width = adapted_frame_->GetWidth();
-        stats.adapted_height = adapted_frame_->GetHeight();
+        stats.adapted_width = static_cast<int>(adapted_frame_->GetWidth());
+        stats.adapted_height = static_cast<int>(adapted_frame_->GetHeight());
       } else {
         stats.adapted_width = stats.adapted_height = -1;
       }
